\name{MyEigen2-package}
\alias{MyEigen2-package}
\alias{MyEigen2}
\docType{package}
\title{
\packageTitle{MyEigen2}
}
\description{
\packageDescription{MyEigen2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MyEigen2}
\packageIndices{MyEigen2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MyEigen2}

Maintainer: \packageMaintainer{MyEigen2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
